+++ gen_saga_test.go
@@ -9,6 +9,27 @@ type testGenSaga struct {
 	GenSaga
 }
 
+func (gs *testGenSaga) InitSaga(p *Process, args ...interface{}) (GenSagaOptions, interface{}) {
+	opts := GenSagaOptions{}
+	return opts, nil
+}
+
+func (gs *testGenSaga) HandleCancel(tx GenSagaTransaction, state GenSagaState) error {
+	return nil
+}
+
+func (gs *testGenSaga) HandleCanceled(tx GenSagaTransaction, reason string, state GenSagaState) error {
+	return nil
+}
+
+func (gs *testGenSaga) HandleDone(tx GenSagaTransaction, result interface{}, state GenSagaState) error {
+	return nil
+}
+
+func (gs *testGenSaga) HandleTimeout(tx GenSagaTransaction, timeout int, state GenSagaState) error {
+	return nil
+}
+
 func TestGenSagaSimple(t *testing.T) {
 	fmt.Printf("\n=== Test GenSagaSimple\n")
 	fmt.Printf("Starting node: nodeGenSagaSimple01@localhost...")